--- frontend/components/Header.tsx
@@ -1,123 +1,0 @@
-import React, { ReactElement, useEffect, useState } from "react";
-import Link from "next/link";
-import clsx from "clsx";
-import { PropTypes } from "@material-ui/core";
-import AppBar from "@material-ui/core/AppBar";
-import Button from "@material-ui/core/Button";
-import Container from "@material-ui/core/Container";
-import Drawer from "@material-ui/core/Drawer";
-import Hidden from "@material-ui/core/Hidden";
-import IconButton from "@material-ui/core/IconButton";
-import Menu from "@material-ui/icons/Menu";
-import Toolbar from "@material-ui/core/Toolbar";
-import Typography from "@material-ui/core/Typography";
-
-import useStyles from "assets/jss/components/header";
-
-type ColorExpanded = PropTypes.Color | "transparent";
-
-interface ChangeColorOnScroll {
-  color: ColorExpanded;
-  height: string | number;
-}
-
-interface HeaderProps {
-  absolute?: string;
-  brand?: string;
-  changeColorOnScroll?: ChangeColorOnScroll;
-  color?: ColorExpanded;
-  fixed?: boolean;
-  rightLinks?: ReactElement;
-}
-
-function Header(props: HeaderProps): ReactElement {
-  const classes = useStyles();
-  const [mobileOpen, setMobileOpen] = useState(false);
-
-  useEffect(() => {
-    if (props.changeColorOnScroll) {
-      window.addEventListener("scroll", headerColorChange);
-    }
-    return function cleanup() {
-      if (props.changeColorOnScroll) {
-        window.removeEventListener("scroll", headerColorChange);
-      }
-    };
-  }, []);
-
-  const handleDrawerToggle = () => {
-    setMobileOpen(!mobileOpen);
-  };
-
-  const headerColorChange = () => {
-    const { color, changeColorOnScroll } = props;
-    const windowsScrollTop = window.pageYOffset;
-    if (windowsScrollTop > changeColorOnScroll.height) {
-      document.body
-        .getElementsByTagName("header")[0]
-        .classList.remove(classes[color]);
-      document.body
-        .getElementsByTagName("header")[0]
-        .classList.add(classes[changeColorOnScroll.color]);
-    } else {
-      document.body
-        .getElementsByTagName("header")[0]
-        .classList.add(classes[color]);
-      document.body
-        .getElementsByTagName("header")[0]
-        .classList.remove(classes[changeColorOnScroll.color]);
-    }
-  };
-
-  const { color, rightLinks, brand, fixed, absolute } = props;
-  return (
-    <AppBar
-      className={clsx({
-        [classes.appBar]: true,
-        [classes[color]]: color,
-        [classes.absolute]: absolute,
-        [classes.fixed]: fixed,
-      })}
-      color={color}>
-      <Container maxWidth="xl">
-        <Toolbar className={classes.container}>
-          <Link href="/">
-            <Button>
-              <Typography
-                className={classes.title}
-                component="div"
-                variant="h4">
-                {brand}
-              </Typography>
-            </Button>
-          </Link>
-          <Hidden smDown implementation="css">
-            {rightLinks}
-          </Hidden>
-          <Hidden mdUp>
-            <IconButton
-              color="inherit"
-              aria-label="open drawer"
-              onClick={handleDrawerToggle}>
-              <Menu />
-            </IconButton>
-          </Hidden>
-        </Toolbar>
-        <Hidden mdUp implementation="css">
-          <Drawer
-            variant="temporary"
-            anchor={"right"}
-            open={mobileOpen}
-            classes={{
-              paper: classes.drawerPaper,
-            }}
-            onClose={handleDrawerToggle}>
-            <div className={classes.appResponsive}>{rightLinks}</div>
-          </Drawer>
-        </Hidden>
-      </Container>
-    </AppBar>
-  );
-}
-
-export default Header;
