+++ frontend/package.json
@@ -23,7 +23,6 @@
     "moment": "2.29.1",
     "next": "11.1.0",
     "peerjs": "1.3.2",
-    "query-string": "7.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-json-view": "1.21.3",
